apps/neogen-jd-generator/requirements.txt
streamlit==1.38.0
pydantic==2.9.2
python-docx==1.1.2
reportlab==4.2.2
Pillow==10.4.0
markdown==3.6
fpdf2==2.7.9